uart-4: fully support ec, print, dtu and kill build warnning

+++ dtu.h
@@ -1,5 +1,5 @@
 #ifndef DTU_H
 #define DTU_H
 void dtu_init(void);
-void dtu_send(int ec, int temp, float ph);
+void dtu_send(int ec, float temp, float ph);
 #endif
