+++ ec.h
@@ -1,5 +1,5 @@
 #ifndef EC_H
 #define EC_H
-void ec_read(int *ec, int *temp);
+void ec_read(int *ec, float *temp);
 void ec_init(void);
 #endif
